Move structured data to head
